feat: imporve node process script

+++ scripts/application_stop.sh
@@ -1,4 +1,13 @@
 #!/bin/bash
 #Stopping existing node servers
 echo "Stopping any existing node servers"
-pkill -f node
+# Get the process ID (PID) of the Node process
+PID=$(pgrep -f "node")
+# Check if the Node process is running
+if [ -n "$PID" ]; then
+  # If the process ID is not empty, kill the Node process
+  echo "Killing node process with PID $PID"
+  pkill -f node
+else
+  echo "Node process is not running"
+fi
